+++ src/main/java/bike/guyona/exdepot/sortingrules/SortingRuleMatcher.java
@@ -1,15 +1,10 @@
 package bike.guyona.exdepot.sortingrules;
 
+import bike.guyona.exdepot.ExDepotMod;
 import bike.guyona.exdepot.capability.StorageConfig;
 import net.minecraft.item.ItemStack;
-import net.minecraft.tileentity.TileEntity;
 
-import java.util.LinkedHashSet;
 import java.util.Set;
-import java.util.Vector;
-
-import static bike.guyona.exdepot.ExDepotMod.proxy;
-import static bike.guyona.exdepot.capability.StorageConfigProvider.STORAGE_CONFIG_CAPABILITY;
 
 public class SortingRuleMatcher {
 //    private Vector<TileEntity> chests;
@@ -36,8 +31,8 @@ public class SortingRuleMatcher {
 //    }
 
     public static boolean matchConfig(ItemStack stack, StorageConfig config) {
-        for (Class<? extends AbstractSortingRule> ruleClass : proxy.sortingRuleProvider.ruleClasses) {
-            AbstractSortingRule itemRule =  proxy.sortingRuleProvider.fromItemStack(stack, ruleClass);
+        for (Class<? extends AbstractSortingRule> ruleClass : ExDepotMod.sortingRuleProvider.ruleClasses) {
+            AbstractSortingRule itemRule =  ExDepotMod.sortingRuleProvider.fromItemStack(stack, ruleClass);
             Set<? extends AbstractSortingRule> sortingRules = config.getRules(ruleClass);
             if (sortingRules == null){
                 continue;
