+++ src/main/java/bike/guyona/exdepot/sortingrules/item/ItemSortingRule.java
@@ -7,7 +7,9 @@ import bike.guyona.exdepot.sortingrules.AbstractSortingRule;
 import net.minecraft.client.Minecraft;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.NBTTagCompound;
+import net.minecraft.nbt.CompoundNBT;
+import net.minecraft.util.ResourceLocation;
+import net.minecraftforge.registries.ForgeRegistries;
 
 import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
@@ -19,11 +21,11 @@ public class ItemSortingRule extends AbstractSortingRule {
 
     private final String itemId;
     private final Integer itemDamage;
-    private final NBTTagCompound itemTags;
+    private final CompoundNBT itemTags;
     private boolean useNbt = true;
     private ItemStack itemCache;
 
-    ItemSortingRule(String itemId, int itemDamage, NBTTagCompound nbt) {
+    ItemSortingRule(String itemId, int itemDamage, CompoundNBT nbt) {
         this.itemId = itemId;
         this.itemDamage = itemDamage;
         this.itemTags = nbt;
@@ -31,18 +33,14 @@ public class ItemSortingRule extends AbstractSortingRule {
 
     ItemSortingRule(ItemStack stack) {
         this.itemId = stack.getItem().getRegistryName().toString();
-        if (stack.getHasSubtypes()) {
-            this.itemDamage = stack.getItemDamage();
-        } else {
-            this.itemDamage = 0;
-        }
+        this.itemDamage = stack.getDamage();
         // TODO Can't do this. There are way too many extra tags that could be added. Look through registry and find the item.
-        itemTags = stack.getTagCompound();
+        itemTags = stack.getTag();
     }
 
     ItemStack getItem() {
         if (itemCache == null) {
-            Item item = Item.getByNameOrId(itemId);
+            Item item = ForgeRegistries.ITEMS.getValue(new ResourceLocation((itemId)));
             if (item != null) {
                 for (ItemStack stack : getSubtypes(item)) {
                     if (matches(stack)) {
@@ -83,8 +81,8 @@ public class ItemSortingRule extends AbstractSortingRule {
         } else if (thing instanceof ItemStack) {
             ItemStack stack = (ItemStack) thing;
             return itemId.equals(stack.getItem().getRegistryName().toString()) &&
-                    (!stack.getHasSubtypes() || itemDamage == stack.getItemDamage()) &&
-                    (!useNbt || itemTags == null || itemTags.equals(stack.getTagCompound()));
+                    itemDamage == stack.getDamage() &&
+                    (!useNbt || itemTags == null || itemTags.equals(stack.getTag()));
         }
         return false;
     }
@@ -94,12 +92,12 @@ public class ItemSortingRule extends AbstractSortingRule {
         if (getItem() == null) {
             return "BROKEN:"+itemId+":"+Integer.toString(itemDamage)+(useNbt ? ":NBT" : "");
         }
-        return getItem().getDisplayName();
+        return getItem().getDisplayName().toString();
     }
 
     @Override
     public void draw(int left, int top, float zLevel) {
-        Minecraft mc = Minecraft.getMinecraft();
+        Minecraft mc = Minecraft.getInstance();
         ItemStack stack = getItem();
         if (stack != null) {
             GuiHelpers.drawItem(left,
