+++ src/main/java/bike/guyona/exdepot/sortingrules/itemcategory/ItemCategorySortingRule.java
@@ -6,7 +6,7 @@ import bike.guyona.exdepot.helpers.GuiHelpers;
 import bike.guyona.exdepot.sortingrules.AbstractSortingRule;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.resources.I18n;
-import net.minecraft.creativetab.CreativeTabs;
+import net.minecraft.item.ItemGroup;
 
 import java.nio.ByteBuffer;
 import java.util.Arrays;
@@ -17,24 +17,24 @@ public class ItemCategorySortingRule extends AbstractSortingRule {
     static final long serialVersionUID = 10;
 
     Integer category;
-    private CreativeTabs categoryCache;
+    private ItemGroup categoryCache;
 
     ItemCategorySortingRule(int tabIndex) {
         this.category = tabIndex;
     }
 
-    ItemCategorySortingRule(CreativeTabs tab) {
+    ItemCategorySortingRule(ItemGroup tab) {
         this.category = AccessHelpers.getTabIndex(tab);
         this.categoryCache = tab;
     }
 
-    CreativeTabs getTab() {
+    ItemGroup getTab() {
         if (categoryCache == null) {
-            for (CreativeTabs tab:CreativeTabs.CREATIVE_TAB_ARRAY) {
+            for (ItemGroup tab:ItemGroup.GROUPS) {
                 if (Arrays.asList(DISALLOWED_CATEGORIES).contains(tab)) {
                     continue;
                 }
-                if (category.equals(tab.getTabIndex())) {
+                if (category.equals(tab.getIndex())) {
                     categoryCache = tab;
                     break;
                 }
@@ -57,23 +57,23 @@ public class ItemCategorySortingRule extends AbstractSortingRule {
     public boolean matches(Object thing) {
         if (thing instanceof ItemCategorySortingRule) {
             return equals(thing);
-        } else if (thing instanceof CreativeTabs) {
-            return category.equals(AccessHelpers.getTabIndex((CreativeTabs) thing));
+        } else if (thing instanceof ItemGroup) {
+            return category.equals(AccessHelpers.getTabIndex((ItemGroup) thing));
         }
         return false;
     }
 
     @Override
     public String getDisplayName() {
-        return I18n.format(getTab().getTranslatedTabLabel());
+        return I18n.format(getTab().getTabLabel());
     }
 
     @Override
     public void draw(int left, int top, float zLevel) {
-        Minecraft mc = Minecraft.getMinecraft();
-        CreativeTabs tab = getTab();
+        Minecraft mc = Minecraft.getInstance();
+        ItemGroup tab = getTab();
         GuiHelpers.drawItem(left,
-                top, tab.getIconItemStack(), mc.fontRenderer);
+                top, tab.getIcon(), mc.fontRenderer);
         mc.fontRenderer.drawString(
                 getDisplayName(),
                 left + StorageConfigGui.ICON_WIDTH,
